Server/flask-kafka/agent2/consumer/requirements.txt
kafka-python
pymongo


